tests: cmd in test def is now full command
The cmdargs parameter in test definitions wera all the arguments to
give to prown command for the test. Now, the cmd parameter is the full
command, the path of the prown binary is set appropriately in the
testing environment through the $BIN$ placeholder.

This way, it becomes possible to prefix the prown command (eg. with
strace for debug purpose).