tests: run test in child with dropped ruid/rgid
The python script launch.py now forks for each test to set its
ruid/rgid to the user defined in the test.

Unfortunately, we cannot just change temporarily the euid as prown
relies on the ruid to identify the current user, in order to correctly
identify the user when prown is installed with setuid 0 (resulting in
euid 0).

As changing ruid/rgid is a one way trip, we must fork to keep a process
with uid 0 for the following tests.