Add manpage
The manpage is written in markdown using pandoc. The corresponding
build rule has been to the makefile.

The manpage is focused on command usage and detailed behaviour in some
corner cases (eg. symbolic link, root of project directories, etc).

The README has been updated to remove usage stuff and redirect to the
manpage. The README is now focused on sysops stuff (purpose,
installation, configuration) and maintainer guidelines.