Add feature to support of group ACL w/ write perm
Prown now accepts users to change owner of files in a project directory
when they are member of a group that has an ACL with write permission
attached to the root this project directory.

Some verbose output have been modified to match the new logic.

Tests are added to test this new feature.