Better output to understand granting logic
This commit just modifies the verbose output so the users can more
easily understand the granting logic behind prown, at runtime.

The code uses getgrgid() in several places, it is not thread safe as it
uses static memory area. It is not a problem for prown as long as it is
not multithreaded.

Tests are updated accordingly.